do not try to apply alpha mask if no alpha channel is present

+++ lib/class.pdf.php
@@ -3968,17 +3968,21 @@ EOT;
       $imagick = new Imagick($file);
       $imagick->setFormat('png');
 
-      // Get opacity channel (negative of alpha channel)
-      $alpha_channel = $imagickClonable ? clone $imagick : $imagick->clone();
-      $alpha_channel->separateImageChannel(Imagick::CHANNEL_ALPHA);
-      $alpha_channel->negateImage(true);
-      $alpha_channel->writeImage($tempfile_alpha);
+        if (0 !== $imagick->getImageAlphaChannel()) {
+            // Get opacity channel (negative of alpha channel)
+            $alpha_channel = $imagickClonable ? clone $imagick : $imagick->clone();
+            $alpha_channel->separateImageChannel(Imagick::CHANNEL_ALPHA);
+            $alpha_channel->negateImage(true);
+            $alpha_channel->writeImage($tempfile_alpha);
 
-      // Cast to 8bit+palette
-      $imgalpha_ = imagecreatefrompng($tempfile_alpha);
-      imagecopy($imgalpha, $imgalpha_, 0, 0, 0, 0, $wpx, $hpx);
-      imagedestroy($imgalpha_);
-      imagepng($imgalpha, $tempfile_alpha);
+            // Cast to 8bit+palette
+            $imgalpha_ = imagecreatefrompng($tempfile_alpha);
+            imagecopy($imgalpha, $imgalpha_, 0, 0, 0, 0, $wpx, $hpx);
+            imagedestroy($imgalpha_);
+            imagepng($imgalpha, $tempfile_alpha);
+        } else {
+            $tempfile_alpha = null;
+        }
 
       // Make opaque image
       $color_channels = new Imagick();
@@ -4037,17 +4041,21 @@ EOT;
       imagepng($imgplain, $tempfile_plain);
     }
 
-    // embed mask image
-    $this->addImagePng($tempfile_alpha, $x, $y, $w, $h, $imgalpha, true);
-    imagedestroy($imgalpha);
+      // embed mask image
+      if ($tempfile_alpha) {
+          $this->addImagePng($tempfile_alpha, $x, $y, $w, $h, $imgalpha, true);
+          imagedestroy($imgalpha);
+      }
 
-    // embed image, masked with previously embedded mask
-    $this->addImagePng($tempfile_plain, $x, $y, $w, $h, $imgplain, false, true);
-    imagedestroy($imgplain);
+      // embed image, masked with previously embedded mask
+      $this->addImagePng($tempfile_plain, $x, $y, $w, $h, $imgplain, false, true);
+      imagedestroy($imgplain);
 
-    // remove temp files
-    unlink($tempfile_alpha);
-    unlink($tempfile_plain);
+      // remove temp files
+      if ($tempfile_alpha) {
+          unlink($tempfile_alpha);
+      }
+      unlink($tempfile_plain);
   }
 
   /**
